Move stacks into core package
