+++ java/cdk/src/main/java/sleeper/cdk/SleeperCdkApp.java
@@ -33,32 +33,17 @@ import sleeper.cdk.stack.CompactionStack;
 import sleeper.cdk.stack.CompactionStatusStoreResources;
 import sleeper.cdk.stack.CompactionStatusStoreStack;
 import sleeper.cdk.stack.DashboardStack;
-import sleeper.cdk.stack.DynamoDBStateStoreStack;
 import sleeper.cdk.stack.GarbageCollectorStack;
 import sleeper.cdk.stack.IngestBatcherStack;
 import sleeper.cdk.stack.IngestStack;
 import sleeper.cdk.stack.IngestStacks;
 import sleeper.cdk.stack.IngestStatusStoreResources;
 import sleeper.cdk.stack.IngestStatusStoreStack;
-import sleeper.cdk.stack.InstanceRolesStack;
 import sleeper.cdk.stack.KeepLambdaWarmStack;
-import sleeper.cdk.stack.LoggingStack;
-import sleeper.cdk.stack.ManagedPoliciesStack;
 import sleeper.cdk.stack.PartitionSplittingStack;
-import sleeper.cdk.stack.PropertiesStack;
 import sleeper.cdk.stack.QueryQueueStack;
 import sleeper.cdk.stack.QueryStack;
-import sleeper.cdk.stack.S3StateStoreStack;
-import sleeper.cdk.stack.StateStoreCommitterStack;
-import sleeper.cdk.stack.StateStoreStacks;
-import sleeper.cdk.stack.TableDataStack;
-import sleeper.cdk.stack.TableIndexStack;
 import sleeper.cdk.stack.TableMetricsStack;
-import sleeper.cdk.stack.TopicStack;
-import sleeper.cdk.stack.TransactionLogSnapshotStack;
-import sleeper.cdk.stack.TransactionLogStateStoreStack;
-import sleeper.cdk.stack.TransactionLogTransactionStack;
-import sleeper.cdk.stack.VpcStack;
 import sleeper.cdk.stack.WebSocketQueryStack;
 import sleeper.cdk.stack.bulkimport.BulkImportBucketStack;
 import sleeper.cdk.stack.bulkimport.CommonEmrBulkImportStack;
@@ -69,6 +54,21 @@ import sleeper.cdk.stack.bulkimport.EmrStudioStack;
 import sleeper.cdk.stack.bulkimport.PersistentEmrBulkImportStack;
 import sleeper.cdk.stack.core.ConfigBucketStack;
 import sleeper.cdk.stack.core.CoreStacks;
+import sleeper.cdk.stack.core.DynamoDBStateStoreStack;
+import sleeper.cdk.stack.core.InstanceRolesStack;
+import sleeper.cdk.stack.core.LoggingStack;
+import sleeper.cdk.stack.core.ManagedPoliciesStack;
+import sleeper.cdk.stack.core.PropertiesStack;
+import sleeper.cdk.stack.core.S3StateStoreStack;
+import sleeper.cdk.stack.core.StateStoreCommitterStack;
+import sleeper.cdk.stack.core.StateStoreStacks;
+import sleeper.cdk.stack.core.TableDataStack;
+import sleeper.cdk.stack.core.TableIndexStack;
+import sleeper.cdk.stack.core.TopicStack;
+import sleeper.cdk.stack.core.TransactionLogSnapshotStack;
+import sleeper.cdk.stack.core.TransactionLogStateStoreStack;
+import sleeper.cdk.stack.core.TransactionLogTransactionStack;
+import sleeper.cdk.stack.core.VpcStack;
 import sleeper.cdk.util.Utils;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.validation.OptionalStack;
