+++ java/cdk/src/main/java/sleeper/cdk/stack/CompactionStatusStoreStack.java
@@ -18,6 +18,7 @@ package sleeper.cdk.stack;
 import software.amazon.awscdk.NestedStack;
 import software.constructs.Construct;
 
+import sleeper.cdk.stack.core.ManagedPoliciesStack;
 import sleeper.core.properties.instance.InstanceProperties;
 
 public class CompactionStatusStoreStack extends NestedStack {
