+++ java/cdk/src/main/java/sleeper/cdk/stack/IngestStatusStoreResources.java
@@ -19,6 +19,7 @@ package sleeper.cdk.stack;
 import software.amazon.awscdk.services.iam.IGrantable;
 import software.constructs.Construct;
 
+import sleeper.cdk.stack.core.ManagedPoliciesStack;
 import sleeper.core.properties.instance.InstanceProperties;
 
 import static sleeper.core.properties.instance.IngestProperty.INGEST_STATUS_STORE_ENABLED;
