+++ java/cdk/src/main/java/sleeper/cdk/stack/IngestStatusStoreStack.java
@@ -19,6 +19,7 @@ package sleeper.cdk.stack;
 import software.amazon.awscdk.NestedStack;
 import software.constructs.Construct;
 
+import sleeper.cdk.stack.core.ManagedPoliciesStack;
 import sleeper.core.properties.instance.InstanceProperties;
 
 public class IngestStatusStoreStack extends NestedStack {
