+++ java/cdk/src/main/java/sleeper/cdk/stack/core/ConfigBucketStack.java
@@ -25,8 +25,6 @@ import software.amazon.awscdk.services.s3.IBucket;
 import software.constructs.Construct;
 
 import sleeper.cdk.jars.BuiltJars;
-import sleeper.cdk.stack.LoggingStack;
-import sleeper.cdk.stack.ManagedPoliciesStack;
 import sleeper.cdk.util.AutoDeleteS3Objects;
 import sleeper.cdk.util.Utils;
 import sleeper.core.properties.instance.InstanceProperties;
