+++ java/cdk/src/main/java/sleeper/cdk/stack/core/StateStoreGrants.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.cdk.stack;
+package sleeper.cdk.stack.core;
 
 import java.util.stream.Stream;
